devtools/server/tests/mochitest/test_memory_attach_02.html

<!DOCTYPE HTML>
<html>
<!--
Bug 960671 - Test attaching and detaching while in the wrong state.
-->
<head>
  <meta charset="utf-8">
  <title>Memory monitoring actor test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="memory-helpers.js" type="application/javascript"></script>
<script>
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  (async function() {
    const { memory, target } = await startServerAndGetSelectedTabMemory();

    let e = null;
    try {
      await memory.detach();
    } catch (ee) {
      e = ee;
    }
    ok(e, "Should have hit the wrongState error");

    await memory.attach();

    e = null;
    try {
      await memory.attach();
    } catch (ee) {
      e = ee;
    }
    ok(e, "Should have hit the wrongState error");

    await memory.detach();
    destroyServerAndFinish(target);
  })();
};
</script>
</pre>
</body>
</html>
